move customeError file to utils
